refoactored code
